Extract firewall header, prior commit was scan and history headers
